# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000" #Api runs locally in port 3000
info:
  title: WASAPhoto
  description: "WASAPhoto API documentation"
  version: 1.0.0
components:
  securitySchemes:  #Security of API, for authentication use Bearer where token represented in JWT Format
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /profile/photo: #4)we are uploading photo that's why it is post
    post:
      tags:
        - Users
      summary: "Uploads or updates the user's profile photo"
      description: "Allows a user to upload a new profile picture."
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: "Profile photo file upload object"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Object containing the profile photo file"
              properties:
                profilePicture:
                  type: string
                  format: binary
                  description: "The profile photo file to upload"
                  minLength: 1
                  maxLength: 256
              required:
                - profilePicture
      responses:
        '200':
          description: "Profile photo updated successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating successful profile photo update"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the update was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid file or update error)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid file format"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
  /profile/username: #5) set new username nickname ,since we edit existing data post
    post:
      tags:
        - Users
      summary: "Set user's username"
      description: "Allows a user to update their username."
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        description: "New username"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request body for updating the user's username"
              properties:
                username:
                  type: string
                  description: "The new username"
                  example: "NewUsername"
                  minLength: 3
                  maxLength: 32
                  pattern: "^[a-zA-Z0-9_]+$"
              required:
                - username
      responses:
        '200':
          description: "Username updated successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating the username was updated successfully"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the username was updated successfully"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /group/photo: #7)update group photo
    post:
      tags:
        - Groups
      summary: "Set group photo"
      description: "Allows an admin to update the group photo."
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: "New group photo file"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Request body for updating the group photo"
              properties:
                groupId:
                  type: integer
                  format: int64
                  description: "ID of the group"
                  example: 123
                groupPhoto:
                  type: string
                  format: binary
                  description: "The new group photo file"
                  minLength: 1
                  maxLength: 256
              required:
                - groupId
                - groupPhoto
      responses:
        '200':
          description: "Group photo updated successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating the group photo was updated"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the operation was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message/uncomment:
    post:
      tags:
        - Messages
      summary: "Remove a comment from a message"
      description: "Allows a user to remove a previously added comment from a message."
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Uncomment request details"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for removing a comment"
              properties:
                commentId:
                  type: integer
                  format: int64
                  description: "ID of the comment to remove"
                  example: 67890
              required:
                - commentId
      responses:
        '200':
          description: "Comment removed successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating the comment was removed"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the comment was successfully removed"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...